Added permissions to the get single attendance admin page

+++ src/pages/api/attendance/admin/[id].ts
@@ -1,26 +1,64 @@
 import { Prisma } from '@prisma/client';
 
-import { attendanceSelectQuery as selectQuery, prisma } from '../../../../db';
-import { auth } from '../../../../middlewares';
+import { permissions } from '../../../../config';
+import {
+	attendanceSelectQuery as selectQuery,
+	prisma,
+	getSingleAttendance,
+} from '../../../../db';
+import { getRecord, getUserObjectPermissions } from '../../../../db/utils';
+import { admin } from '../../../../middlewares';
 import { AttendanceCreateType } from '../../../../types';
+import { hasModelPermission } from '../../../../utils';
+import { NextApiErrorMessage } from '../../../../utils/classes';
 import { attendanceCreateSchema } from '../../../../validators';
 
-export default auth()
+export default admin()
 	.get(async (req, res) => {
-		const data = await prisma.attendance.findUnique({
-			where: {
-				id: req.query.id as string,
+		const record = await getRecord({
+			model: 'attendance',
+			perm: 'attendance',
+			permission: 'VIEW',
+			objectId: req.query.id as string,
+			user: req.user,
+			getData() {
+				return getSingleAttendance(req.query.id as string);
 			},
-			select: selectQuery,
 		});
+
+		if (!record) throw new NextApiErrorMessage(403);
+
+		if (!record.data)
+			return res.status(404).json({
+				status: 'error',
+				message: 'Attendance record with the specified ID does not exist!',
+			});
 
 		return res.status(200).json({
 			status: 'success',
 			message: 'Fetched attendance!',
-			data,
+			data: record.data,
 		});
 	})
 	.put(async (req, res) => {
+		let hasPerm =
+			req.user.isSuperUser ||
+			hasModelPermission(req.user.allPermissions, [
+				permissions.attendance.EDIT,
+			]);
+
+		if (!hasPerm) {
+			const perms = await getUserObjectPermissions({
+				modelName: 'attendance',
+				permission: 'EDIT',
+				objectId: req.query.id as string,
+				userId: req.user.id,
+			});
+			hasPerm = perms.edit;
+		}
+
+		if (!hasPerm) throw new NextApiErrorMessage(403);
+
 		const data: AttendanceCreateType =
 			await attendanceCreateSchema.validateAsync({ ...req.body });
 
@@ -79,6 +117,24 @@ export default auth()
 		});
 	})
 	.delete(async (req, res) => {
+		let hasPerm =
+			req.user.isSuperUser ||
+			hasModelPermission(req.user.allPermissions, [
+				permissions.attendance.DELETE,
+			]);
+
+		if (!hasPerm) {
+			const perms = await getUserObjectPermissions({
+				modelName: 'attendance',
+				permission: 'DELETE',
+				objectId: req.query.id as string,
+				userId: req.user.id,
+			});
+			hasPerm = perms.delete;
+		}
+
+		if (!hasPerm) throw new NextApiErrorMessage(403);
+
 		await prisma.attendance.delete({
 			where: {
 				id: req.query.id as string,
